first push

+++ src/main/java/cs455/overlay/wireformats/Protocol.java
@@ -2,7 +2,7 @@ package cs455.overlay.wireformats;
 
 public interface Protocol {
     int REGISTER_REQ = 1;
-    byte REGISTER_RES = 2;
-    byte DEREGISTER_REQ = 3;
-    byte DEREGISTER_RES = 4;
+    int REGISTER_RES = 2;
+    int DEREGISTER_REQ = 3;
+    int DEREGISTER_RES = 4;
 }
